Organize JRuby console implementations
Group implementations by backend and isolate platform-specific
constants.
