+++ io-console.gemspec
@@ -39,14 +39,14 @@ Gem::Specification.new do |s|
     s.require_paths.unshift('jruby/lib')
     s.files.concat(%w[
       jruby/lib/io/console.rb
-      jruby/lib/io/console/bsd_console.rb
+      jruby/lib/io/console/backend/ffi.rb
+      jruby/lib/io/console/backend/stty.rb
+      jruby/lib/io/console/backend/stub.rb
+      jruby/lib/io/console/backend/windows.rb
       jruby/lib/io/console/common.rb
-      jruby/lib/io/console/linux_console.rb
-      jruby/lib/io/console/native_console.rb
-      jruby/lib/io/console/stty_console.rb
-      jruby/lib/io/console/stub_console.rb
-      jruby/lib/io/console/windows_constants.rb
-      jruby/lib/io/console/windows_console.rb
+      jruby/lib/io/console/constants/bsd.rb
+      jruby/lib/io/console/constants/linux.rb
+      jruby/lib/io/console/constants/windows.rb
       jruby/lib/io/console/version.rb
     ])
   end
