+++ jruby/lib/io/console/backend/ffi.rb
