+++ jruby/lib/io/console/backend/stub.rb
