+++ jruby/lib/io/console/backend/windows.rb
