+++ jruby/lib/io/console/constants/bsd.rb
@@ -170,5 +170,3 @@ module IO::Console::LibC
   attach_function :tcflush, [ :int, :int ], :int
   attach_function :ioctl, [ :int, :ulong, :varargs ], :int
 end
-
-require_relative 'native_console'
