+++ jruby/lib/io/console/constants/windows.rb
